requirements.txt
opencv-python==4.8.0
pillow==10.0.0
pyautogui==0.9.54
plyer==2.1.0
python-dotenv==1.0.0
requests==2.31.0
line-bot-sdk==3.5.0
psutil==5.9.5
watchdog==3.0.0
scapy==2.5.0
pyinstaller==6.5.0
pystray==0.19.5